+++ scripts/spell-darkness.js
@@ -1,0 +1,13 @@
+const version = "0.1.0";
+const optionName = "Darkness Spell";
+
+try {
+	
+	if (args[0].macroPass === "templatePlaced") {
+		//const { actor, token, lArgs } = MidiMacros.targets(args);
+		
+	}
+
+} catch (err) {
+    console.error(`${optionName} : ${version}`, err);
+}