+++ scripts/spell-thunderous-smite.js
@@ -1,0 +1,87 @@
+const version = "0.1.0";
+const optionName = "Thunderous Smite";
+const squaresPushed = 2;
+
+try {
+	if (args[0].macroPass === "DamageBonus") {	
+		let workflow = MidiQOL.Workflow.getWorkflow(args[0].uuid);
+		let actor = workflow?.actor;
+		const target = args[0].hitTargets[0];
+		let tactor = target?.actor;
+		const ttoken = canvas.tokens.get(args[0].hitTargets[0].object.id);
+		let pusher = canvas.tokens.get(args[0].tokenId);
+
+		// validate targeting
+		if (!actor || !target) {
+		  console.log(`${optionName}: no target selected`);
+		  return {};
+		}
+
+		// make sure it's an allowed attack
+		const at = args[0].item?.data?.actionType;
+		if (!at || !["mwak"].includes(at)) {
+			console.log(`${optionName}: not an eligible attack: ${at}`);
+			return {};
+		}
+		
+		// TODO check target type is a creature ???
+		
+		// remove the effect, since it is one-time
+		let effect = actor.effects?.find(i=>i.data.label === optionName);
+		let conc = actor.effects?.find(i=>i.data.label === game.i18n.localize("Concentrating"));
+		if (effect) {
+			await MidiQOL.socket().executeAsGM("removeEffects", { actorUuid: actor.uuid, effects: [effect.id] });
+			if (conc) await MidiQOL.socket().executeAsGM("removeEffects", { actorUuid: actor.uuid, effects: [conc.id] });
+		}
+
+		// push the target logic
+		const ability = actor.data.data.attributes.spellcasting;
+		const abilityBonus = actor.data.data.abilities[ability].mod;
+		const dc = 8 + actor.data.data.attributes.prof + abilityBonus;
+		let saveType = game.i18n.localize("str");
+		let save = await MidiQOL.socket().executeAsGM("rollAbility", { request: "save", targetUuid: tactor.uuid, ability: saveType, 
+			options: { chatMessage: true, fastForward: false } });
+			
+	    if (save.total < dc) {
+			await game.dfreds?.effectInterface.addEffect({ effectName: 'Prone', uuid: target.actor.uuid });
+			await wait(300);
+			await pushTarget(pusher, ttoken);
+		}
+
+		// add damage bonus
+		const diceMult = args[0].isCritical ? 4: 2;
+		let damageType = game.i18n.localize("thunder");
+		return {damageRoll: `${diceMult}d6[${damageType}]`, flavor: optionName};
+	}
+
+
+} catch (err) {
+    console.error(`${optionName}:  ${version}`, err);
+}
+
+async function wait(ms) { return new Promise(resolve => { setTimeout(resolve, ms); }); }
+
+async function pushTarget(sourceToken, targetToken) {
+	let newCenter = getAllowedPushLocation(sourceToken, targetToken, squaresPushed);
+	if(!newCenter) {
+		return ui.notifications.error(`${optionName} - no room to push ${targetToken.name}`);
+	}
+	const tobj = targetToken.document.object;
+	newCenter = canvas.grid.getSnappedPosition(newCenter.x - tobj.w / 2, newCenter.y - tobj.h / 2, 1);
+	const mutationData = { token: {x: newCenter.x, y: newCenter.y}};
+	await warpgate.mutate(targetToken.document, mutationData, {}, {permanent: true});
+}
+
+function getAllowedPushLocation(sourceToken, targetToken, maxSquares) {
+	for (let i = maxSquares; i > 0; i--) {
+		const knockbackPixels = i * canvas.grid.size;
+		const ray = new Ray(sourceToken.center, targetToken.center);
+		const newCenter = ray.project((ray.distance + knockbackPixels)/ray.distance);
+		const isAllowedLocation = canvas.sight.testVisibility({x: newCenter.x, y: newCenter.y}, {object: targetToken.Object});
+		if(isAllowedLocation) {
+			return newCenter;
+		}
+	}
+	
+	return null;
+}